Function documentation almost completed

+++ output.h
@@ -6,13 +6,13 @@
 
 /// @brief Function prints underscore which are borders of the each row in the board 
 /// @param width -> The width of the rows (number of entries in row)
-void underscore(int width);
+void underscore(unsigned long long width);
 
 /// @brief Function prints the board (with borders) and all pieces in board in current state
 /// @param height -> The height of the printed board
 /// @param width  -> The width of the printed board
 /// @param board -> The state of the board to be printed
-void printArray(int height, int width,char board[][width], player p1, player p2);
+void printArray(unsigned long long height, unsigned long long width,char board[][width], player p1, player p2);
 
 /// @brief Function resets colour to default color
 void reset_color();
